Abstract out Deck as a component
